Restructured code"
