+++ api/app-v2.js
@@ -1,0 +1,6 @@
+import request from 'request';
+
+const appV2 = (req, res) => {
+}
+
+export default appV2;
